Move devfile related exceptions to a separate package
