+++ wsmaster/che-core-api-devfile/src/main/java/org/eclipse/che/api/devfile/server/DevfileRecipeFormatException.java
@@ -11,6 +11,8 @@
  */
 package org.eclipse.che.api.devfile.server;
 
+import org.eclipse.che.api.devfile.server.exception.DevfileException;
+
 /** Thrown when the provided content of recipe-type tool is empty or invalid. */
 public class DevfileRecipeFormatException extends DevfileException {
