+++ wsmaster/che-core-api-devfile/src/test/java/org/eclipse/che/api/devfile/server/convert/CommandConverterTest.java
@@ -21,8 +21,8 @@ import com.google.common.collect.ImmutableMap;
 import java.util.HashMap;
 import org.eclipse.che.api.devfile.model.Action;
 import org.eclipse.che.api.devfile.model.Command;
-import org.eclipse.che.api.devfile.server.DevfileFormatException;
-import org.eclipse.che.api.devfile.server.WorkspaceExportException;
+import org.eclipse.che.api.devfile.server.exception.DevfileFormatException;
+import org.eclipse.che.api.devfile.server.exception.WorkspaceExportException;
 import org.eclipse.che.api.workspace.server.model.impl.CommandImpl;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
